pyproject.toml
[project]
name = "climate-forecasting"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    {name = "Rob Ellis", email = "robandrewells@gmail.com"},
]
requires-python = ">=3.12"
dependencies = [
    "pandas>=2.2.3",
    "plotly>=5.24.1",
    "scipy>=1.14.1",
    "statsmodels>=0.14.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "hydra-core>=1.3.2",
    "pre-commit>=4.0.1",
    "pydantic>=2.10.4",
    "ruff>=0.8.4",
]

[tool.hatch.build]
packages = ["src/climate_forecasting"]

[tool.ruff]
line-length = 79

[tool.ruff.lint]
select = ["D101", "D102", "D103"]

[tool.ruff.lint.pydocstyle]
convention = "google"
